Alteração no design e tipo da variavel price

+++ cardapio/src/main/java/com/example/cardapio/dtos/FoodResponseDTO.java
@@ -2,5 +2,5 @@ package com.example.cardapio.dtos;
 
 import java.util.UUID;
 
-public record FoodResponseDTO(String tittle, String image, Integer price) {
+public record FoodResponseDTO(String tittle, String image, Double price) {
 }
